sign up auth with firebase working and adding user to sqlite3
